rename `pages` folder to `organisms`
