+++ web/src/app/page.tsx
@@ -1,4 +1,4 @@
-import { LandingPage } from "@/pages/LandingPage";
+import { LandingPage } from "@/organisms/LandingPage";
 
 export default function Page() {
   return (
